Make pivot data

Make pivot data

pivot_data(rawdata_grouped, cols_to_ignore = "unit", key_col = "model",
  value_col = "quantity_sum")

Arguments

rawdata_grouped
grouped rawdata as retrieved by function group_data()

cols_to_ignore
column names not to be considered for transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "unit")

key_col
column name containing the key for transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "model")

value_col
column name containing the values that will be used during transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "quantity_sum")

Value

data.frame with the aggregated results for all different"lci_method"

Examples

zipfile <- system.file("extdata/Beispiel_Auswertung.zip", package = "kwb.lca")
temp <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = temp)
rawdata <- kwb.lca::import_rawdata(csv_dir = temp)

#> [1] "Importing csv file 'C:\\Users\\mrustl.KWB\\AppData\\Local\\Temp\\RtmpCQDrGb/Beispiel_Auswertung/Ref_High_C+high N.csv'"

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> [1] "Importing csv file 'C:\\Users\\mrustl.KWB\\AppData\\Local\\Temp\\RtmpCQDrGb/Beispiel_Auswertung/Ref_High_C+N.csv'"

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> [1] "Importing csv file 'C:\\Users\\mrustl.KWB\\AppData\\Local\\Temp\\RtmpCQDrGb/Beispiel_Auswertung/Ref_Low_C+high N.csv'"

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> [1] "Importing csv file 'C:\\Users\\mrustl.KWB\\AppData\\Local\\Temp\\RtmpCQDrGb/Beispiel_Auswertung/Ref_Low_C+N.csv'"

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

data_grouped <- kwb.lca::group_data(rawdata)
data_pivot <- kwb.lca::pivot_data(data_grouped)
head(data_pivot)

#> # A tibble: 6 x 6
#> # Groups:   lci_method, process [6]
#>   lci_method      process    `0a_Ref_Low_C+N` `0b_Ref_High_C+~ `0c_Ref_Low_C+h~
#>   <chr>           <chr>                 <dbl>            <dbl>            <dbl>
#> 1 cumulative ene~ T11: Elec~          145350.          135761.          162218.
#> 2 cumulative ene~ T12: Elec~           30204.           29674.           30950.
#> 3 cumulative ene~ T13: FeCl3           17734.           23087.           25930.
#> 4 cumulative ene~ T14: Poly~           21443.           22178.           21871.
#> 5 cumulative ene~ T15: Slud~           22137.           22794.           23126.
#> 6 cumulative ene~ T21: Mono~            5687.            5829.            5906.
#> # ... with 1 more variable: `0d_Ref_High_C+high N` <dbl>